requirements.txt
torch
torchvision
numpy
scipy
matplotlib
pillow
